Add files via upload
I want to make ray box intersections like in the volumetric clouds example.
I've been struggling with an annoying problem for a long time. I made a snippet from my app.
You can move with wasd, the arrow keys and space.
My problem are the ugly white dashed edges. I just can't figure out where it's coming from and what I can do about it.
I have tried to reduce the app as much as possible in the hope of limiting the phenomenon.
The RayBox intersection can be found in the shader cloudsFS.js in the shader folder in resources. threejs is initialized in threejs-component.js and in the main the composer, renderer and camera are passed to cloud. There's nothing that would be complicated in any way, but as I said, I just can't figure out where these ugly lines are coming from